+++ src/commands/package/version/create/list.ts
@@ -1,8 +1,17 @@
 /*
- * Copyright (c) 2022, salesforce.com, inc.
- * All rights reserved.
- * Licensed under the BSD 3-Clause license.
- * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
+ * Copyright 2025, Salesforce, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
 import { Flags, loglevel, orgApiVersionFlagWithDeprecations, SfCommand } from '@salesforce/sf-plugins-core';
@@ -94,7 +103,15 @@ export class PackageVersionCreateListCommand extends SfCommand<CreateListCommand
       this.table({ data, overflow: 'wrap', title: chalk.blue(`Package Version Create Requests  [${results.length}]`) });
     }
 
-    return results;
+    // Filter out unwanted fields from JSON output
+    // TotalNumberOfMetadataFiles and TotalSizeOfMetadataFiles are intentionally excluded from display
+    const filteredResults = results.map((r) => {
+      // eslint-disable-next-line @typescript-eslint/no-unused-vars
+      const { TotalNumberOfMetadataFiles, TotalSizeOfMetadataFiles, ...filteredResult } = r;
+      return filteredResult;
+    });
+
+    return filteredResults as CreateListCommandResult;
   }
 
   // Queries Package2Version for the name and version number of the packages and adds that data