+++ src/commands/package/version/retrieve.ts
@@ -1,8 +1,17 @@
 /*
- * Copyright (c) 2022, salesforce.com, inc.
- * All rights reserved.
- * Licensed under the BSD 3-Clause license.
- * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
+ * Copyright 2025, Salesforce, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
 import path from 'node:path';
@@ -10,7 +19,7 @@ import {
   Flags,
   loglevel,
   orgApiVersionFlagWithDeprecations,
-  requiredOrgFlagWithDeprecations,
+  requiredHubFlagWithDeprecations,
   SfCommand,
 } from '@salesforce/sf-plugins-core';
 import { Messages } from '@salesforce/core/messages';
@@ -36,7 +45,7 @@ export class PackageVersionRetrieveCommand extends SfCommand<PackageVersionRetri
   public static readonly flags = {
     loglevel,
     'api-version': orgApiVersionFlagWithDeprecations,
-    'target-org': requiredOrgFlagWithDeprecations,
+    'target-dev-hub': requiredHubFlagWithDeprecations,
     package: Flags.string({
       char: 'p',
       summary: messages.getMessage('flags.package.summary'),
@@ -51,7 +60,7 @@ export class PackageVersionRetrieveCommand extends SfCommand<PackageVersionRetri
 
   public async run(): Promise<PackageVersionRetrieveCommandResult> {
     const { flags } = await this.parse(PackageVersionRetrieveCommand);
-    const connection = flags['target-org'].getConnection(flags['api-version']);
+    const connection = flags['target-dev-hub'].getConnection(flags['api-version']);
     const options = {
       subscriberPackageVersionId: flags.package ?? '',
       destinationFolder: flags['output-dir'],