+++ test/commands/package/versionPromoteUpdate.nut.ts
@@ -1,8 +1,17 @@
 /*
- * Copyright (c) 2022, salesforce.com, inc.
- * All rights reserved.
- * Licensed under the BSD 3-Clause license.
- * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
+ * Copyright 2025, Salesforce, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
 import path from 'node:path';
@@ -57,7 +66,9 @@ describe('package:version:promote / package:version:update', () => {
       'CreatedBy',
       'ConvertedFromVersionId',
       'CodeCoverage',
-      'VersionNumber'
+      'VersionNumber',
+      'TotalNumberOfMetadataFiles',
+      'TotalSizeOfMetadataFiles'
     );
     expect(result?.Id).to.match(/08c.{15}/);
     expect(result?.Package2Id).to.match(/0Ho.{15}/);